requirement.txt
requests==2.27.1
python-dotenv==0.19.2
pandas==1.4.1
websockets==10.2 
websocket-client==1.3.1
prometheus-client==0.13.1
Flask==2.0.3
asyncio==3.4.3